Move `ModalTopLevelComponent(Union)` to `components` package
