+++ src/main/java/net/dv8tion/jda/api/components/Component.java
@@ -24,7 +24,6 @@ import net.dv8tion.jda.api.components.container.ContainerChildComponent;
 import net.dv8tion.jda.api.components.section.SectionAccessoryComponent;
 import net.dv8tion.jda.api.components.section.SectionContentComponent;
 import net.dv8tion.jda.api.interactions.modals.Modal;
-import net.dv8tion.jda.api.interactions.modals.ModalTopLevelComponent;
 
 import javax.annotation.CheckReturnValue;
 import javax.annotation.Nonnull;
