+++ src/main/java/net/dv8tion/jda/api/components/ModalTopLevelComponentUnion.java
@@ -14,11 +14,8 @@
  * limitations under the License.
  */
 
-package net.dv8tion.jda.api.interactions.modals;
+package net.dv8tion.jda.api.components;
 
-import net.dv8tion.jda.api.components.Component;
-import net.dv8tion.jda.api.components.IComponentUnion;
-import net.dv8tion.jda.api.components.UnknownComponent;
 import net.dv8tion.jda.api.components.actionrow.ActionRow;
 
 import javax.annotation.Nonnull;
