+++ src/main/java/net/dv8tion/jda/internal/components/tree/ModalComponentTreeImpl.java
@@ -16,9 +16,9 @@
 
 package net.dv8tion.jda.internal.components.tree;
 
+import net.dv8tion.jda.api.components.ModalTopLevelComponent;
+import net.dv8tion.jda.api.components.ModalTopLevelComponentUnion;
 import net.dv8tion.jda.api.components.replacer.ComponentReplacer;
-import net.dv8tion.jda.api.interactions.modals.ModalTopLevelComponent;
-import net.dv8tion.jda.api.interactions.modals.ModalTopLevelComponentUnion;
 import net.dv8tion.jda.api.interactions.modals.tree.ModalComponentTree;
 import net.dv8tion.jda.internal.components.utils.ComponentsUtil;
 import net.dv8tion.jda.internal.utils.Checks;
