+++ src/main/java/net/dv8tion/jda/internal/interactions/modal/ModalImpl.java
@@ -17,8 +17,8 @@
 package net.dv8tion.jda.internal.interactions.modal;
 
 import net.dv8tion.jda.api.components.Components;
+import net.dv8tion.jda.api.components.ModalTopLevelComponentUnion;
 import net.dv8tion.jda.api.interactions.modals.Modal;
-import net.dv8tion.jda.api.interactions.modals.ModalTopLevelComponentUnion;
 import net.dv8tion.jda.api.utils.data.DataArray;
 import net.dv8tion.jda.api.utils.data.DataObject;
 import net.dv8tion.jda.internal.components.AbstractComponentImpl;
